feat(trigger): schedule weekly organize digest via trigger.dev cron
Adds organizeWeeklyCronTask (schedules.task, Monday 8am UTC) that queues
an organize run per user with triggeredBy: cron, which is what makes
send-organize-complete.ts send the weekly digest email. Previously this
only fired if something external called the /organize/run cron endpoint,
which nothing in the repo actually did.

Extracted the shared insertRunOrSkip/runOrganizeForAllUsers logic out of
the orpc handler so both the endpoint and the new schedule use one
implementation.

Closes #304
