+++ apps/api/src/trigger/organize.ts
@@ -10,6 +10,7 @@ export { sendWaitlistApprovedEmailTask } from "@harmonia/common/trigger/tasks/em
 export { sendWaitlistConfirmationEmailTask } from "@harmonia/common/trigger/tasks/emails/send-waitlist-confirmation";
 export { sendWelcomeEmailTask } from "@harmonia/common/trigger/tasks/emails/send-welcome";
 export { organizePipeline } from "@harmonia/common/trigger/tasks/organize";
+export { organizeWeeklyCronTask } from "@harmonia/common/trigger/tasks/organize-weekly-cron";
 export { refreshLibrarySnapshotsTask } from "@harmonia/common/trigger/tasks/spotify/refresh-library-snapshots";
 export { syncUserLibraryTask } from "@harmonia/common/trigger/tasks/spotify/sync-user-library";
 export { artistsStageTask } from "@harmonia/common/trigger/tasks/stages/artists";
